minimum viable base
